pyproject.toml
[project]
name = "z-alpha-risk-management"
version = "0.1.0"
description = "Z-Alpha Securities"
authors = [
    {name = "maggyy666",email = "119632961+maggyy666@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "ibapi (>=9.81.1.post1,<10.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "psycopg2-binary (>=2.9.9,<3.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pandas (>=2.0.0,<3.0.0)",
    "python-multipart (>=0.0.6,<1.0.0)",
    "scipy (>=1.12.0,<2.0.0)",
    "arch (>=7.2.0,<8.0.0)",
    "yfinance (>=0.2.65,<0.3.0)",
    "bcrypt (>=4.0,<6.0)",
    "pyjwt (>=2.8.0,<3.0.0)",
    "pydantic-settings (>=2.5.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
name = "z-alpha-securities"
version = "0.1.0"
description = "Z-Alpha Securities Risk Management Platform"
authors = ["maggyy666 <your.email@example.com>"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"
